requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
crispy-bootstrap5==2024.10
crispy-tailwind==1.0.3
DateTime==5.5
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.6
django-bootstrap4==24.4
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-environ==0.11.2
django-heroku==0.3.1
django-js-asset==3.0.1
django-taggit==6.1.0
django-user-visit==2.0
djangorestframework==3.15.2
filelock==3.16.1
gunicorn==23.0.0
MouseInfo==0.1.3
mysql==0.0.3
mysqlclient==2.2.5
packaging==24.2
pathlib==1.0.1
pillow==11.0.0
pipenv==2024.4.1
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
setuptools==75.8.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
virtualenv==20.27.1
wheel==0.45.1
whitenoise==6.8.2
zope.interface==7.2
